Add factory class to create ModelOutput instances
This class is necessary for models to be able to easily return
ModelOutput instances without knowing any of the details about what
they're returning. The intended flow of use is as follows.

A config file gets passed into `build_model`, which then decides the
model architecture and instantiates a ModelOutputFactory, say `factory`,
with the appropriate kwargs depending on the config. Then the model
itself is instantiated and recieves `factory` so
that in its `forward` method, all it needs to do is call
`factory.create_output(y)`, where `y` is the raw tensor output.